\name{catnames}
\alias{catnames}
\title{Get all category names (titles).}
\usage{
  catnames()
}
\value{
  A character vector of category names
}
\description{
  Get all category names (titles).
}
\examples{
category.names <- catnames()
}
\author{
  Joona Lehtomäki <joona.lehtomaki@gmail.com>
}
\keyword{avoindata,}
\keyword{categories}

