\name{count_answers}
\alias{count_answers}
\title{Get the amount of answers per day.}
\usage{
  count_answers()
}
\value{
  A DataFrame. Each row has a count and date column.
}
\description{
  Get the amount of answers per day.
}
\examples{
na <- count_answers()
}
\author{
  Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
  response2df nquestions
}
\keyword{answers}
\keyword{avoindata,}
\keyword{categories,}

