\name{count_questions}
\alias{count_questions}
\title{Get the amount of questions per day.}
\usage{
  count_questions()
}
\value{
  A DataFrame. Each row has a count and date column.
}
\description{
  Get the amount of questions per day.
}
\examples{
nq <- count_questions()
#'
}
\author{
  Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
  response2df nanswers
}
\keyword{avoindata,}
\keyword{categories,}
\keyword{questions}

