\name{gettags}
\alias{gettags}
\title{Get all the tags, counts, and wordids.}
\usage{
  gettags()
}
\value{
  A list of tag items. Each item has attributes title,
  count, and wordid
}
\description{
  Get all the tags, counts, and wordids.
}
\examples{
tags <- get_tags()
}
\author{
  Joona Lehtomäki <joona.lehtomaki@gmail.com>
}
\keyword{avoindata,}
\keyword{tags}

