\name{nanswers}
\alias{nanswers}
\title{Get the amount of answers per day.}
\usage{
  nanswers()
}
\value{
  A DataFrame. Each row has a count and date column.
}
\description{
  Get the amount of answers per day.
}
\examples{
nq <- nanswers()
}
\author{
  Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
  response2df nquestions
}
\keyword{answers}
\keyword{avoindata,}
\keyword{categories,}

