\name{nquestions}
\alias{nquestions}
\title{Get the amount of quesitions per day.}
\usage{
  nquestions()
}
\value{
  A DataFrame. Each row has a count and date column.
}
\description{
  Get the amount of quesitions per day.
}
\examples{
nq <- nquestions()
}
\author{
  Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\keyword{avoindata,}
\keyword{categories,}
\keyword{questions}

