\name{taginfo}
\alias{taginfo}
\title{Get all the tags, counts, and wordids.}
\usage{
  taginfo()
}
\value{
  A DataFrame with each row containg columns title, count,
  and wordid
}
\description{
  Get all the tags, counts, and wordids.
}
\note{
  Should the results be cached?
}
\examples{
tags <- taginfo()
}
\author{
  Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
  response2df
}
\keyword{avoindata,}
\keyword{tags}

