\name{tagnames}
\alias{tagnames}
\title{Get all tag names (titles).}
\usage{
  tagnames()
}
\value{
  A character vector of tag names
}
\description{
  Get all tag names (titles).
}
\examples{
tags <- tagnames()
}
\author{
  Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\keyword{avoindata,}
\keyword{tags}

