\name{gettags}
\alias{gettags}
\title{Get all the tags, counts, and wordids.}
\usage{
  gettags()
}
\value{
  A DataFrame with each row containg columns title, count,
  and wordid
}
\description{
  Get all the tags, counts, and wordids.
}
\note{
  Should the results be cached?
}
\examples{
tags <- get_tags()
}
\author{
  Joona Lehtomäki <joona.lehtomaki@gmail.com>
}
\keyword{avoindata,}
\keyword{tags}

